Create Count odd even
In the previous module we learnt about the basics of CPP.  Now, we'll move to more complex stuff. Here, we'll learn about arrays.

Given an array A of N elements. The task is to count number of even and odd elements in the array.

Input:
First line of input contains number of testcases T. For each testcase, there will be two line of input, first line of which contains N and next line contains N array elements.

Output:
For each testcase, print the number of odd elements first and then number of even elements seperated by space in a single line.